requirements.txt
streamlit
requests
pillow
openai
